+++ Scripts/Spells/Necromancy/TransformationSpell.cs
@@ -15,7 +15,7 @@ namespace Server.Spells.Necromancy
         public virtual int PoisResistOffset => 0;
         public virtual int NrgyResistOffset => 0;
         public override bool BlockedByHorrificBeast => false;
-        public virtual double TickRate => 1.0;
+        public virtual int TickRate => 1;
 
         public override bool CheckCast()
         {
